Clinic Records API (unsolved)/requirements.txt
asgiref==3.11.0
Django==4.2.11
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==25.0
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
uritemplate==4.2.0
